requirements.txt
# Core Flask Framework
Flask==3.0.0
Werkzeug==3.0.1

# CORS Support for Android
Flask-CORS==4.0.0

# Compression (gzip/brotli)
Flask-Compress==1.14
Brotli==1.1.0

# Rate Limiting
Flask-Limiter==3.5.0

# Authentication & Security
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
cryptography==41.0.7
Flask-HTTPAuth==4.8.0

# Environment Configuration
python-dotenv==1.0.0

# Database & ORM
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9  # PostgreSQL adapter
Flask-Migrate==4.0.5

# Caching
Flask-Caching==2.1.0
redis==5.0.1

# Data Processing (existing dependencies)
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2  # Excel file support

# Validation
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
email-validator==2.1.0

# Logging & Monitoring
structlog==23.2.0
python-json-logger==2.0.7

# HTTP Client (for FCM)
requests==2.31.0
httpx==0.25.2

# Firebase Cloud Messaging
firebase-admin==6.3.0

# File Upload Optimization
Flask-Uploads==0.2.1

# API Documentation
flask-swagger-ui==4.11.1
apispec==6.3.1
apispec-webframeworks==1.0.0

# Testing (development)
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
locust==2.20.0  # Load testing

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
